2025_DEVELOPER_WEEK/NavigatingTheEdgeCloudBridge/demos/1-thin-client/ws/requirements.txt
# batch
mediapipe==0.10.14
opencv-python==4.10.0.84
numpy==1.26.4
xgboost==2.1.1
hyperopt==0.2.7

# train
# torch==2.4.0
torch==2.5.1
# torchvision==0.19.0
scikit-learn==1.6.0
pandas==2.2.3
av==12.3.0
# fastai==2.7.16
shap==0.46.0

# other?
# tqdm==4.66.4
tqdm==4.67.1

# stt and tts
deepgram-sdk==3.7.7
# pyaudio==0.2.14

# llm
openai==1.57.2

# flask
Flask==3.0.3
